Finalizada parte Autorização

+++ app/Http/Controllers/ProjectController.php
@@ -36,7 +36,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
     
     /**
@@ -58,7 +58,11 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
-        return $this->repository->with('user')->with('client')->find($id);
+        if($this->checkProjectPermission($id) == false){
+            return ['error' => "Access Forbidden"];
+        }
+
+        return $this->repository->with('owner')->with('client')->with('members')->find($id);
     }
 
     /**
@@ -70,6 +74,10 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+       if($this->checkProjectPermission($id) == false){
+           return ['error' => "Access Forbidden"];
+       }
+
        return $this->service->update($request->all(),$id);
     }
 
@@ -81,6 +89,31 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if($this->checkProjectOwner($id) == false){
+            return ['error' => "Access Forbidden"];
+        }
+
         return $this->repository->delete($id);
     }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermission($projectId)
+    {
+        if($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)){
+            return true;
+        }
+        return false;
+    }
+
 }
